requirements.txt
argparse==1.4.0
openai==0.27.4
pinecone-client==2.2.1
PyYAML==6.0